pyproject.toml
[project]
name = "latent-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chardet>=7.4.3",
    "datasets>=4.8.4",
    "decord>=0.6.0",
    "diffusers>=0.37.1",
    "einops>=0.8.2",
    "multiprocess>=0.70.19",
    "opencv-python>=4.13.0.92",
    "progress-table>=3.2.2",
    "psutil>=7.2.2",
    "pyqt6>=6.11.0",
    "pyturbojpeg>=2.2.0",
    "ruamel-yaml>=0.19.1",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "timm>=1.0.26",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.6.2",
    "wandb>=0.26.1",
]
